+++ react-native/Components/mulecules/CardMaior/index.js
@@ -1,10 +1,12 @@
 import { CardContainer, CardImage } from "./style";
 import React from "react";
+import { useNavigation } from "@react-navigation/native";
 
 
 export const CardMaior = ({item}) =>{
+    const navigation = useNavigation();
     return(
-        <CardContainer>
+        <CardContainer onPress={()=> navigation.navigate('noticia')}>
             <CardImage source={{uri: item.image_url}}/>
         </CardContainer>
     );
