+++ react-native/Components/organism/HomeList/styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components/native'
+
+export const ListContainer = styled.View`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: flex-start;
+  width: 100%;
+`
