+++ react-native/package.json
@@ -13,6 +13,7 @@
     "@expo-google-fonts/roboto": "^0.2.2",
     "@react-navigation/bottom-tabs": "^6.3.1",
     "@react-navigation/native": "^6.0.10",
+    "@react-navigation/stack": "^6.2.1",
     "expo": "~45.0.0",
     "expo-app-loading": "~2.0.0",
     "expo-font": "~10.1.0",
@@ -24,6 +25,7 @@
     "react-dom": "17.0.2",
     "react-native": "0.68.2",
     "react-native-drop-shadow": "^0.0.4",
+    "react-native-gesture-handler": "^2.4.2",
     "react-native-paper": "^4.12.1",
     "react-native-safe-area-context": "^4.2.5",
     "react-native-screens": "^3.13.1",
